% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lastval}
\alias{lastval}
\title{lastval
This function returns the vlaue at the last index of a list}
\usage{
lastval(vec)
}
\arguments{
\item{vec}{A list that you want to obtain the last value from}
}
\value{
Returns the last value of a list
}
\description{
lastval
This function returns the vlaue at the last index of a list
}
\examples{
lastval(c(2,3,4))
}
